[Data Views] Extract pure type definitions to @kbn/data-views-types package
## Summary

This PR addresses part of #129433 by extracting pure type definitions from the
`data_views` plugin into a new `@kbn/data-views-types` package.

### Changes

- Created new `@kbn/data-views-types` package containing:
  - `DataViewSpec`, `DataViewAttributes`, `DataViewType`
  - `FieldSpec`, `FieldFormatMap`, `FieldAttrs`, `DataViewFieldMap`
  - `RuntimeField`, `RuntimeFieldSpec`, `RuntimeType`
  - Constants: `RUNTIME_FIELD_TYPES`, `META_FIELDS`, `DATA_VIEW_SAVED_OBJECT_TYPE`

- Updated `data_views` plugin to re-export from the new package for backward compatibility

- Cleaned up stale comments referencing deprecated `IIndexPattern` type

### Notes

- `DataViewBase` and `DataViewFieldBase` remain in `@kbn/es-query` as they serve an
  architectural purpose (minimal interfaces for query building)
- `IIndexPattern` and `IFieldType` have already been fully migrated away

Fixes #129433
