+++ src/platform/packages/shared/kbn-data-views-types/kibana.jsonc
@@ -1,0 +1,5 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/data-views-types",
+  "owner": "@elastic/kibana-data-discovery"
+}
