+++ src/platform/packages/shared/kbn-data-views-types/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@kbn/data-views-types",
+  "private": true,
+  "version": "1.0.0",
+  "license": "Elastic License 2.0 OR AGPL-3.0-only OR SSPL-1.0",
+  "main": "./index",
+  "types": "./index"
+}
