+++ src/platform/plugins/shared/kql/public/autocomplete/providers/kql_query_suggestion/field.tsx
@@ -38,7 +38,6 @@ export const setupGetFieldSuggestions: KqlQuerySuggestionProvider<QuerySuggestio
       indexPatterns.map((indexPattern) => {
         return indexPattern.fields.filter(isFilterable);
       })
-      // temp until IIndexPattern => DataView
     ) as DataViewField[];
     const search = `${prefix}${suffix}`.trim().toLowerCase();
     const matchingFields = allFields.filter((field) => {
